split up constraints_handler.* into its components
