+++ core/CMakeLists.txt
@@ -13,6 +13,7 @@ target_sources(ginkgo
     base/perturbation.cpp
     base/version.cpp
     constraints/constraints_handler.cpp
+    constraints/zero_rows.cpp
     distributed/partition.cpp
     factorization/ic.cpp
     factorization/ic.cpp
