+++ cuda/CMakeLists.txt
@@ -75,7 +75,7 @@ target_sources(ginkgo_cuda
     base/version.cpp
     components/device_matrix_data_kernels.cu
     components/prefix_sum_kernels.cu
-    constraints/constraints_handler_kernels.cu
+    constraints/utility_kernels.cu
     distributed/partition_kernels.cu
     factorization/factorization_kernels.cu
     factorization/ic_kernels.cu
