+++ cuda/constraints/utility_kernels.cu
@@ -30,7 +30,7 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************<GINKGO LICENSE>*******************************/
 
-#include "core/constraints/constraints_handler_kernels.hpp"
+#include "core/constraints/utility_kernels.hpp"
 
 
 #include <memory>
