+++ hip/CMakeLists.txt
@@ -160,7 +160,7 @@ set(GINKGO_HIP_SOURCES
     base/version.hip.cpp
     components/device_matrix_data_kernels.hip.cpp
     components/prefix_sum_kernels.hip.cpp
-    constraints/constraints_handler_kernels.hip.cpp
+    constraints/utility_kernels.hip.cpp
     distributed/partition_kernels.hip.cpp
     factorization/factorization_kernels.hip.cpp
     factorization/ic_kernels.hip.cpp
