+++ omp/CMakeLists.txt
@@ -7,7 +7,7 @@ target_sources(ginkgo_omp
     base/version.cpp
     components/device_matrix_data_kernels.cpp
     components/prefix_sum_kernels.cpp
-    constraints/constraints_handler_kernels.cpp
+    constraints/utility_kernels.cpp
     distributed/partition_kernels.cpp
     factorization/factorization_kernels.cpp
     factorization/ic_kernels.cpp
