+++ reference/CMakeLists.txt
@@ -9,7 +9,7 @@ target_sources(ginkgo_reference
     components/reduce_array_kernels.cpp
     components/precision_conversion_kernels.cpp
     components/prefix_sum_kernels.cpp
-    constraints/constraints_handler_kernels.cpp
+    constraints/utility_kernels.cpp
     distributed/partition_kernels.cpp
     factorization/factorization_kernels.cpp
     factorization/ic_kernels.cpp
